requirements.txt
mss
opencv-python
pyautogui
pillow
pyyaml
pytesseract
